+++ {{ cookiecutter.project_name }}/project/settings/development.py
@@ -1,9 +1,3 @@
-"""
-=======================
-Development environment
-=======================
-
-"""
 from configurations import values
 
 from .base import ComposedProjectSettings