+++ {{ cookiecutter.project_name }}/project/settings/production.py
@@ -1,9 +1,3 @@
-"""
-======================
-Production environment
-======================
-
-"""
 from configurations import values
 
 from .base import ComposedProjectSettings
@@ -11,7 +5,7 @@ from .base import ComposedProjectSettings
 
 class Production(ComposedProjectSettings):
     """
-    Settings for kubernetes environment.
+    Settings for production environment.
     """
     # Labelize the deployed production environment name
     ENVIRONMENT = values.Value("Production", environ_name="DEPLOYED_ENVIRONMENT")