+++ {{ cookiecutter.project_name }}/project/settings/test.py
@@ -1,20 +1,20 @@
-"""
-================
-Test environment
-================
-
-"""
 from configurations import values
 
-from .development import Development
+from .base import ComposedProjectSettings
 
 
-class Test(Development):
+class Test(ComposedProjectSettings):
     """
     Settings for test environment.
 
     Intended to be used only by test runner.
     """
+    DEBUG = True
+
+    # Don't send any email for real, just push them to the shell output
+    EMAIL_BACKEND = values.Value("django.core.mail.backends.console.EmailBackend",
+                                 environ_name="DJANGO_EMAIL_BACKEND")
+
     @property
     def MEDIA_ROOT(self):
         """
@@ -26,6 +26,13 @@ class Test(Development):
     @classmethod
     def post_setup(cls):
         super().post_setup()
+
+        # Disable webpack cache
+        cls.WEBPACK_LOADER["DEFAULT"]["CACHE"] = values.BooleanValue(
+            False,
+            environ_name="WEBPACK_CACHE",
+            environ_prefix=None
+        )
 
         # Patch database setting to use database in memory instead of file
         cls.DATABASES["default"]["NAME"] = values.Value(
@@ -40,3 +47,7 @@ class Test(Development):
                 (k, v)
                 for k, v in cls.TEST_PAGE_TEMPLATES.items()
             ]
+
+        # Force disabling error about Recaptcha API development keys
+        if "captcha.recaptcha_test_key_error" not in cls.SILENCED_SYSTEM_CHECKS:
+            cls.SILENCED_SYSTEM_CHECKS.append("captcha.recaptcha_test_key_error")