+++ docs/history.rst
@@ -9,6 +9,8 @@ Version 0.3.3 - Unreleased
 
 * Disabled ``check-migrations`` task from ``quality`` task until CMS plugins have been
   updated for proper Django>=4.0 support, see issue #21 for details;
+* Test environment settings no longer inherit from Development, instead some of
+  Development have been copied to the Test settings;
 
 
 Version 0.3.2 - 2023/01/30